PR/30823: Dave Huang: Panic reading files larger than 4GB on NTFS

+++ sys/fs/ntfs/ntfs_subr.h
@@ -1,4 +1,4 @@
-/*	$NetBSD: ntfs_subr.h,v 1.2 2005/05/29 21:00:29 christos Exp $	*/
+/*	$NetBSD: ntfs_subr.h,v 1.3 2005/07/25 00:48:22 christos Exp $	*/
 
 /*-
  * Copyright (c) 1998, 1999 Semen Ustimenko
@@ -45,8 +45,8 @@ struct ntvattr {
 
 	u_int32_t		va_compression;
 	u_int32_t		va_compressalg;
-	u_int32_t		va_datalen;
-	u_int32_t		va_allocated;
+	u_int64_t		va_datalen;
+	u_int64_t		va_allocated;
 	cn_t	 		va_vcnstart;
 	cn_t	 		va_vcnend;
 	u_int16_t		va_index;
